requirements.txt
google-generativeai
streamlit
google-genai
python-dotenv